backend/requirements.txt
fastapi
uvicorn
firebase-admin
python-multipart
torch
torchvision
captum
pillow
numpy
python-dotenv
opencv-python
